commit de segurança para o app de tarefas

+++ pyutilidades/todo/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render
 from rest_framework.views import APIView
 from rest_framework.response import Response
+from rest_framework.authtoken.models import Token
+from django.contrib.auth.models import User
 from .serializer import Tarefa_serializer,Tarefa
 from django.http import Http404
 from rest_framework import status
@@ -14,15 +16,14 @@ class Tarefa_View(APIView):
         tarefas = Tarefa.objects.all()
         serializer = Tarefa_serializer(tarefas, many=True)
         return Response(serializer.data, status=200)
-    
-    def post(self,request):
-        serializer = Tarefa_serializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data,status=201)
 
 class Detail_Tarefa_View(APIView):
     
+    def verificar_token(self,req):
+        token = req.headers['Authorization'][6:]
+        user = User.objects.get(pk=req.data['user'])
+        return token == Token.objects.get(user=user).key        
+
     def get_object(self, pk):
         try:
             tarefa = Tarefa.objects.filter(user=pk)
@@ -30,21 +31,34 @@ class Detail_Tarefa_View(APIView):
         except Tarefa.DoesNotExist:
             raise Http404
 
+    def post(self,request,pk):
+        if self.verificar_token(request):
+            serializer = Tarefa_serializer(data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(serializer.data,status=201)
+        return Response(data={'token_error':'invalid token'}, status=401)
+
     def get(self, request, pk):
-        tarefa = self.get_object(pk)
-        serializer = Tarefa_serializer(instance=tarefa,many=True)
-        return Response(data=serializer.data, status=status.HTTP_200_OK)
+        if self.verificar_token(request):
+            tarefa = self.get_object(pk)
+            serializer = Tarefa_serializer(instance=tarefa,many=True)
+            return Response(data=serializer.data, status=status.HTTP_200_OK)
+        return Response(data={'token_error':'invalid token'}, status=401)
 
     def delete(self, request, pk):
-        tarefa = self.get_object(pk)
-        tarefa.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if self.verificar_token(request):
+            tarefa = self.get_object(pk)
+            tarefa.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(data={'token_error':'invalid token'}, status=401)
 
     def put(self, request, pk):
-        tarefa = self.get_object(pk)
-        serializer = Tarefa_serializer(instance=tarefa, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        else:
+        if self.verificar_token(request):
+            tarefa = self.get_object(pk)
+            serializer = Tarefa_serializer(instance=tarefa, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(data={'token_error':'invalid token'}, status=401)
